examples/python/aws-cdk/pyproject.toml
[tool.poetry]
name = "cdk-organizer-aws-example"
version = "1.0.0"
description = "AWS CDK Organizer Python Example"
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "templates"

  [[tool.poetry.packages]]
  include = "stacks"

  [tool.poetry.dependencies]
  python = ">=3.8,<3.10"
  "aws-cdk-lib" = "2.39.0"
  "constructs" = "10.1.96"

  [tool.poetry.dependencies.cdk-organizer]
  path = "../../../packages/python"
  develop = true
  extras = ["aws"]

[build-system]
requires = ["poetry-core==1.1.0"]
build-backend = "poetry.core.masonry.api"
